#56 Code refactor
