+++ lib/client.js
@@ -369,7 +369,7 @@
 	var request = safeRequire('request');
 
 	var parseUrl = safeRequire('url', 'parse', function(url) {
-		const link = window.document.createElement('a');
+		var link = window.document.createElement('a');
 		link.setAttribute('href', url);
 
 		return link;
@@ -847,12 +847,12 @@
 			throw new Error('option.serviceVersion is required');
 		}
 
-		// var path = '/socket.io';
-
 		var validationParams = options.validationParams || null;
-		// TODO: error handling if url is not a valid url and no host:
 
 		var parsedUrl = parseUrl(url);
+		if (!parsedUrl.host) {
+			throw new Error('Invalid url and/or host: ' + url);
+		}
 		this.socket = SocketIO(parsedUrl.protocol + '//' + parsedUrl.host);
 		this._attachEvents(validationParams, options.serviceName, options.serviceVersion);
 
@@ -876,11 +876,7 @@
 	SocketIoTransport.prototype.send = function(message, callback) {
 		var self = this;
 
-		if (!self._contextAcknowledged) {
-			// TODO: throw error
-		}
-
-		if (!this.socket || this.socket.disconnected) {
+		if (!this.socket || this.socket.disconnected || !self._contextAcknowledged) {
 			setTimeout(function () {
 				self.send(message, callback);
 			}, 50);
@@ -947,7 +943,7 @@
 		});
 
 		this.socket.on('connect', function () {
-			self.socket.emit('rpc.sio.setContext', {
+			self.socket.emit('rpc.sio.setConnectionContext', {
 				validationParams: validationParams,
 				serviceName: serviceName,
 				serviceVersion: serviceVersion
