+++ lib/transport/base-transport.js
@@ -27,6 +27,40 @@ class BaseTransport extends EventEmitter {
 	sendMessage(/*msg, context, format*/) {
 	}
 
+	onConnect(context) {
+		this.trace.connect(context);
+		this.emit('connect', context);
+	}
+
+	onDisconnect(context) {
+		this.trace.disconnect(context);
+		this.emit('disconnect', context);
+
+		// Un-subscribe events:
+		for (const eventContextMapEntry of this.eventContextMap.entries()) {
+			const eventContextMapKey = eventContextMapEntry[0];
+			const eventContextSet = eventContextMapEntry[1];
+
+			if (eventContextSet.has(context)) {
+				eventContextSet.delete(context);
+
+				if (eventContextSet.size === 0) {
+					const eventListener = this.eventListeners.get(eventContextMapKey);
+					const service = this.eventContextServices.get(eventContextMapKey);
+
+					if (eventListener) {
+						const eventName = eventContextMapKey.substr(eventContextMapKey.indexOf(':') + 1);
+						service.removeListener(eventName, eventListener);
+					}
+
+					this.eventContextMap.delete(eventContextMapKey);
+					this.eventListeners.delete(eventContextMapKey);
+					this.eventContextServices.delete(eventContextMapKey);
+				}
+			}
+		}
+	}
+
 	validateMessage(service, methodName, options, callback) {
 		if (service.requestValidator) {
 			const promise = service.requestValidator(methodName, options, callback);
