+++ lib/transport/socketio-transport.js
@@ -28,11 +28,6 @@ class SocketIOTransport extends BaseTransport {
 		});
 	}
 
-	onConnect(context) {
-		this.trace.connect(context);
-		this.emit('connect', context);
-	}
-
 	sendMessage(msg, context/*, format*/) {
 		if (!(context && context.socket && context.socket.connected)) {
 			return;
@@ -79,8 +74,12 @@ class SocketIOTransport extends BaseTransport {
 
 		this.onConnect(socket);
 
-		socket.on('rpc.sio.setContext', ({validationParams, serviceName, serviceVersion}, cb) => {
-			// TODO: validate the existence of serviceName and serviceVersion
+		socket.on('rpc.sio.setConnectionContext', ({validationParams, serviceName, serviceVersion}, cb) => {
+			if (!(serviceName && serviceVersion)) {
+				const errMsg = "No 'serviceName' or 'serviceVersion' provided to socket.io implementation.";
+				cb(errMsg);
+				throw new Error(errMsg);
+			}
 
 
 			const service = this.registry.getService(`/${serviceName}/${serviceVersion}`);
@@ -96,9 +95,9 @@ class SocketIOTransport extends BaseTransport {
 
 		socket.on('message', (message, cb) => {
 			if (!connectionContext.service) {
-				// TODO: Report meaningful error, maybe also to the client?
-				console.error('No context set yet', message);
-				return;
+				const errMsg = `No connection context set yet ${message}.`;
+				cb(errMsg);
+				throw new Error(errMsg);
 			}
 
 			try {
