+++ lib/transport/ws-transport.js
@@ -7,7 +7,8 @@
 const stream = require('stream');
 const WebSocket = require('ws');
 const jsonrpc = require('./json-rpc');
-const WebSocketBaseTransport = require('./web-socket-base-transport');
+const BaseTransport = require('./base-transport');
+const pathToRegExp = require('path-to-regexp');
 
 const WebSocketServer = WebSocket.Server;
 
@@ -24,7 +25,7 @@ function abortConnection(socket, code, name) {
 	}
 }
 
-class WebSocketTransport extends WebSocketBaseTransport {
+class WebSocketTransport extends BaseTransport {
 	static get type() {
 		return 'WebSocket';
 	}
@@ -35,6 +36,14 @@ class WebSocketTransport extends WebSocketBaseTransport {
 		this.wsServer = null;
 		this.connectionContexts = new WeakMap();
 		this.nextConnectionId = 0;
+
+		this.registry.httpServer.on('connection', socket => {
+			this.onConnect(socket);
+			socket.on('close', () => {
+				this.onDisconnect(socket);
+			});
+		});
+
 		this._setupHandlers();
 	}
 
@@ -161,6 +170,40 @@ class WebSocketTransport extends WebSocketBaseTransport {
 		});
 
 		ws.send('{}');
+	}
+
+	/**
+	 * @param url
+	 * @returns {*}
+	 * @private
+	 */
+	_getServiceAndParams(url) {
+		for (const route of this.registry.routes) {
+			const routeParser = pathToRegExp(`${route}/:_service/:_version`);
+			const execResult = routeParser.exec(url);
+
+			if (!execResult) {
+				continue;
+			}
+
+			const result = routeParser.keys.reduce(function(current, paramKey, paramIndex) {
+				current[paramKey.name] = execResult[paramIndex + 1];
+
+				return current;
+			}, {});
+
+			const service = this.registry.getService(`/${result._service}/${result._version}`);
+			if (service) {
+				delete result._service;
+				delete result._version;
+				return {
+					params: result,
+					service
+				};
+			}
+		}
+
+		return null;
 	}
 }
